Added new tiktok url support

+++ main.py
@@ -12,6 +12,7 @@ MATCH1 = os.getenv('MATCH1', '')
 MATCH2 = os.getenv('MATCH2', '')
 MATCH3 = os.getenv('MATCH3', '')
 MATCH4 = os.getenv('MATCH4', '')
+MATCH5 = os.getenv('MATCH5', '')
 REPLACE = os.getenv('REPLACE', '')
 REPLACE2 = os.getenv('REPLACE2', '')
 REPLACE3 = os.getenv('REPLACE3', '')
@@ -31,6 +32,7 @@ async def on_message(message: discord.Message) -> None:
     instagram_link = re.findall('https://www\.instagram\.com/p/[a-zA-Z0-9_-]+/?(\?[^/]+)?', message.content)
     instagram_reel_link = re.findall('https://www\.instagram\.com/reel/[a-zA-Z0-9_-]+/?(\?[^/]+)?', message.content)
     tiktok_link = re.findall('https:\/\/www\.tiktok\.com\/@[\w\.]+\/video\/\d+', message.content)
+    tiktok_vm_link = re.findall('https:\/\/vm\.tiktok\.com\/[a-zA-Z0-9]+\/?)', message.content)
 
     if twitter_link:
         logger.info(f'{message.guild.name}: {message.author} {message.content}')
@@ -93,6 +95,22 @@ async def on_message(message: discord.Message) -> None:
         else:
             await message.channel.send(new_message, allowed_mentions=allowed_mentions) 
             
+        if DELETE_OP == 1:
+            await message.delete()
+
+    if tiktok_vm_link:
+        logger.info(f'{message.guild.name}: {message.author} {message.content}')
+        new_message = f'{message.author.mention} {PREAMBLE}{message.content.replace(MATCH5, REPLACE3)}'
+        allowed_mentions = discord.AllowedMentions(
+            everyone=message.mention_everyone,
+            users=message.mentions,
+            roles=message.role_mentions
+        )
+        if REPLY_TO == 1:
+            await message.channel.send(new_message, allowed_mentions=allowed_mentions, reference=message)
+        else:
+            await message.channel.send(new_message, allowed_mentions=allowed_mentions)
+
         if DELETE_OP == 1:
             await message.delete()
